+++ src/main/java/com/calculadora/CalculadoraApp/service/OperationService.java
@@ -2,8 +2,8 @@ package com.calculadora.CalculadoraApp.service;
 
 public interface OperationService {
 
-    public Integer addition(Integer numberA, Integer numberB);
+    public Long addition(Long numberA, Long numberB);
 
-    public Integer subtraction(Integer numberA, Integer numberB);
+    public Long subtraction(Long numberA, Long numberB);
 
 }
